Create setup flavors file
