+++ .fvm/flutter_sdk
@@ -1,0 +1,1 @@
+/home/lucas/fvm/versions/2.8.1
