+++ .fvm/fvm_config.json
@@ -1,0 +1,4 @@
+{
+  "flutterSdkVersion": "2.8.1",
+  "flavors": {}
+}
