+++ lib/src/utils/authentication.dart
@@ -1,0 +1,25 @@
+import 'package:shared_preferences/shared_preferences.dart';
+
+class Authentication {
+  static String _tokenKey = "token";
+
+  static Future<bool> authenticated() async {
+    final prefs = await SharedPreferences.getInstance();
+    return prefs.getString(_tokenKey) != null;
+  }
+
+  static void logout() async {
+    final prefs = await SharedPreferences.getInstance();
+    prefs.remove(_tokenKey);
+  }
+
+  static void saveToken(String token) async {
+    final prefs = await SharedPreferences.getInstance();
+    prefs.setString(_tokenKey, token);
+  }
+
+  static Future<String> getToken() async {
+    final prefs = await SharedPreferences.getInstance();
+    return prefs.getString(_tokenKey)!;
+  }
+}
